refactor: keep access key record type local
